Update instantiation and documentation for changing settings
The documentation was less than helpful! Better explained now, and
there's an easier way of accessing the settings object - just
instantiate tinyAutocomplete again.

There's also a small fix for the domains, but they're actually sort of
broken and should be moved to GitHub pages. Let's do that when we get
the time, eh?